backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.6
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.7.4
Pillow>=10.0.0
tqdm>=4.65.0
pytest>=7.0.0
httpx>=0.24.0
